% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{array_split}
\alias{array_split}
\title{Split large data into smaller chunks}
\usage{
array_split(E, number_of_chunks)
}
\arguments{
\item{E}{Count matrix}

\item{number_of_chunks}{Number of chunks to divide the data}
}
\value{
list where each element is a chunk of the original matrix
}
\description{
Split large data into smaller chunks
}
