% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.AddManualTrack}
\alias{CID.AddManualTrack}
\title{Run DEG analysis with Seurat wrapper}
\usage{
CID.AddManualTrack(spring.dir, fn = "categorical_coloring_data.json")
}
\arguments{
\item{spring.dir}{Path to "categorical_coloring_data.json" from SPRING pipeline.}

\item{fn}{Filename, default is 'categorical_coloring_data.json'.}
}
\value{
An additional list element called "Manual_Annotation" added to the file "fn" in spring.dir
}
\description{
Run DEG analysis with Seurat wrapper
}
