% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.append}
\alias{CID.append}
\title{Calculates ICT scores}
\usage{
CID.append(expression, featuresG)
}
\arguments{
\item{expression}{An expression matrix with features (genes) in rows and samples (cells) in columns.}

\item{featuresG}{A list of features, where each list entry is a character vector of markers.}
}
\value{
A matrix of CellID scores; features by cells.
}
\description{
Calculates ICT scores
}
