% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.BatchMode}
\alias{CID.BatchMode}
\title{runs CID.CellID in batch mode}
\usage{
CID.BatchMode(E, f, pval, deep_dive, edges, impute)
}
\arguments{
\item{E}{A list of expression matrices with features (genes) in rows and samples (cells) in columns.}
}
\value{
A list where each element is a CID.CellID result. (See ?CID.CellID)
}
\description{
runs CID.CellID in batch mode
}
