% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.CellID_Seurat}
\alias{CID.CellID_Seurat}
\title{Main function}
\usage{
CID.CellID_Seurat(E, full.dataset = NULL, normalize = T, pval = 0.05,
  data.dir = NULL, entropy = T, louvain = T, omit = NULL,
  sorted = F)
}
\arguments{
\item{E}{A Seurat object}

\item{full.dataset}{If E was subsetted or imputed, full.dataset is the full expression matrix, which is loaded for the detection of novel cell types. Default is NULL.}

\item{normalize}{Normalizes count matrix to the mean library size. Default is TRUE.}

\item{pval}{p-value cutoff for feature selection, as described in the manuscript + markdown file. Default is pval = 0.05.}

\item{data.dir}{directory of SPRING files "edges.csv" and "categorical_coloring_data.json"}

\item{entropy}{cells amended to high entropy labels with respect to their neighbors in the KNN graph are appended "Other" if entropy = TRUE. Default is TRUE.}

\item{louvain}{Louvain community detection is performed, and then used together with Shannon entropy to detect potential novel cell types / states. Default is TRUE.}

\item{omit}{Force remove specific cell types / states with omit. Default is NULL.}

\item{sorted}{If cells are expected to be pure or mostly homogeneous (e.g., by FACs sorting), set sorted = TRUE. Default is FALSE.}
}
\value{
Filtered markers where each marker must have at least ncells that express at least ncounts
}
\description{
Main function
}
