% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Chunk}
\alias{CID.Chunk}
\title{Chunk a dataset}
\usage{
CID.Chunk(E, chunk.dir, number_of_chunks = 30)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame).}

\item{chunk.dir}{A directory where the chunked matrices will be stored. If the directory does not exist, it will be created.}

\item{number_of_chunks}{the number of chunks (e.g., the number of sub-sampled matrices after chunking)}
}
\value{
A directory with chunked matrix
}
\description{
Chunk a dataset
}
