% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.deepdive}
\alias{CID.deepdive}
\title{Deep cell states}
\usage{
CID.deepdive(XX, YY, acOut3, expression, sorted = sorted)
}
\arguments{
\item{XX}{filtered features for CID.append}

\item{YY}{cell type where we are going deeper}

\item{acOut3}{Vector of cell type labels}

\item{expression}{the count matrix}

\item{sorted}{see ?CID.CellID}
}
\value{
Vector for the deep cell state labels
}
\description{
Deep cell states
}
