% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.filter}
\alias{CID.filter}
\title{filters the geneset markers}
\usage{
CID.filter(expression, markersG, pval = 0.1)
}
\arguments{
\item{expression}{An expression matrix with features (genes) in rows and samples (cells) in columns.}

\item{markersG}{A data frame with four columns ('HUGO Symbol', 'Cell Population', 'ENTREZ ID', 'Polarity'). Default is internally set to data(markers_v4).}

\item{pval}{p-value cutoff, as described in the manuscript. Default is 0.1.}
}
\value{
A list of markers and features.
}
\description{
filters the geneset markers
}
