% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.filter2}
\alias{CID.filter2}
\title{filters the geneset markers}
\usage{
CID.filter2(D, markersG, pval = pval, sorted = sorted)
}
\arguments{
\item{D}{An expression matrix with features (genes) in rows and samples (cells) in columns.}

\item{markersG}{A data frame with four columns ('HUGO Symbol', 'Cell Population', 'ENTREZ ID', 'Polarity'). Default is internally set to data(markers_v4).}

\item{pval}{p-value cutoff, as described in the manuscript. Default is 0.01.}

\item{sorted}{see ?CID.CellID}
}
\value{
A list of markers and features.
}
\description{
filters the geneset markers
}
