% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.GetDistMat}
\alias{CID.GetDistMat}
\title{Get distance matrix}
\usage{
CID.GetDistMat(edges, n = 4)
}
\arguments{
\item{edges}{a data frame with two columns; V1 and V2, specifying the cell-cell edge list for the network}

\item{n}{maximum network distance to subtend (n neighbors)}
}
\value{
adjacency matrices for distances < n
}
\description{
Get distance matrix
}
