% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.GetLearningMarkers}
\alias{CID.GetLearningMarkers}
\title{Run DEG analysis with Seurat wrapper}
\usage{
CID.GetLearningMarkers(
  E,
  full.dataset,
  logfc.threshold = 1,
  pseudocount.use = 1
)
}
\arguments{
\item{E}{Expression matrix with genes for rows, samples for columns}
}
\value{
List where each element contains the DEG tables for the one vs. all comparison
}
\description{
Run DEG analysis with Seurat wrapper
}
