% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.GetTrainingSet2}
\alias{CID.GetTrainingSet2}
\title{Get indices of training markers}
\usage{
CID.GetTrainingSet2(E, cr, data.dir = NULL)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{data.dir}{if default, uses the standard Signac markers (see ?CID.SeeMarkers).}
}
\value{
a list of cell indices and names to use for training set
}
\description{
Get indices of training markers
}
