% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Impute}
\alias{CID.Impute}
\title{Main function}
\usage{
CID.Impute(E, do.par = FALSE)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{do.par}{Boolean. If true, imputation is performed in parallel on half of the machines available cores. Default = FALSE.}
}
\value{
imputed expression matrix with only marker genes in rows.
}
\description{
Main function
}
