% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.IsUnique}
\alias{CID.IsUnique}
\title{Extract unique elements}
\usage{
CID.IsUnique(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
boolean, unique elements are TRUE
}
\description{
Extract unique elements
}
