% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.KCore}
\alias{CID.KCore}
\title{K-core calculation on second degree connections}
\usage{
CID.KCore(dM2)
}
\arguments{
\item{dM2}{Adjacency matrix of second degree interactions}
}
\value{
vector of the k-core of each node
}
\description{
K-core calculation on second degree connections
}
