% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Learn}
\alias{CID.Learn}
\title{Learn new cell type markers}
\usage{
CID.Learn(E, cells.ind, refererence = "default")
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{cells.ind}{a named list of the indices of the cells, and the name of the putatitive label}

\item{reference}{if default, uses the standard Signac markers (see ?CID.SeeMarkers).}
}
\value{
an updated reference set with the new cell type
}
\description{
Learn new cell type markers
}
