% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Learn4}
\alias{CID.Learn4}
\title{Learn new cell type markers}
\usage{
CID.Learn4(E)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}
}
\value{
an updated reference set with the new cell type
}
\description{
Learn new cell type markers
}
