% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.LoadEdges}
\alias{CID.LoadEdges}
\title{Load edges from edge list}
\usage{
CID.LoadEdges(data.dir)
}
\arguments{
\item{data.dir}{A directory where "edges.csv" is located}
}
\value{
The edgelist in data frame format
}
\description{
Load edges from edge list
}
