% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.LoadImputedDataFolder}
\alias{CID.LoadImputedDataFolder}
\title{Load chunked imputed data files from directory}
\usage{
CID.LoadImputedDataFolder(chunk.dir)
}
\arguments{
\item{chunk.dir}{As defined in ?CID.chunk}
}
\value{
A sparse matrix with rownames equivalent to the names in genes.txt
}
\description{
Load chunked imputed data files from directory
}
