% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Louvain}
\alias{CID.Louvain}
\title{Community substructure by Louvain community detection}
\usage{
CID.Louvain(edges)
}
\arguments{
\item{edges}{A data frame or matrix with edges}
}
\value{
The community substructures of the graph
}
\description{
Community substructure by Louvain community detection
}
