% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Normalize}
\alias{CID.Normalize}
\title{Library size normalize}
\usage{
CID.Normalize(E)
}
\arguments{
\item{E}{Expression matrix}
}
\value{
Normalized expression matrix to mean of total counts
}
\description{
Library size normalize
}
