% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.PosMarkers}
\alias{CID.PosMarkers}
\title{Run DEG analysis with Seurat wrapper}
\usage{
CID.PosMarkers(E, acn)
}
\arguments{
\item{E}{Expression matrix with genes for rows, samples for columns}

\item{acn}{Vector of cell type labels}
}
\value{
List where each element contains the DEG tables for the one vs. all comparison
}
\description{
Run DEG analysis with Seurat wrapper
}
