% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.PosMarkers2}
\alias{CID.PosMarkers2}
\title{Get positive genes by z-score}
\usage{
CID.PosMarkers2(D, acn)
}
\arguments{
\item{D}{Expression matrix with genes for rows, samples for columns}

\item{acn}{Vector of cell type labels}
}
\value{
List where each element contains the z score for the one vs. all comparison
}
\description{
Get positive genes by z-score
}
