% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Read10Xh5}
\alias{CID.Read10Xh5}
\title{Load 10X count matrix from an h5 file}
\usage{
CID.Read10Xh5(filename, use.names = TRUE)
}
\arguments{
\item{filename}{Directory and filename of the h5 file}

\item{use.names}{Boolean TRUE, see ?Seurat::Read10X_h5}
}
\value{
Count matrix with genes and barcodes
}
\description{
Load 10X count matrix from an h5 file
}
