% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.smooth}
\alias{CID.smooth}
\title{Smoothing function}
\usage{
CID.smooth(ac, dM)
}
\arguments{
\item{ac}{List containing a character vector where each element is a cell type or cell state assignment}

\item{dM}{Distance matrix (see ?CID.GetDistMat)}
}
\value{
Smoothed cell type or cell state assignments
}
\description{
Smoothing function
}
