% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.UnNormalize}
\alias{CID.UnNormalize}
\title{Library size UN-normalize}
\usage{
CID.UnNormalize(E, total_counts)
}
\arguments{
\item{E}{Expression matrix}
}
\value{
Normalized expression matrix to mean of total counts
}
\description{
Library size UN-normalize
}
