% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.UnsupervisedLearning}
\alias{CID.UnsupervisedLearning}
\title{Unsupervised main function}
\usage{
CID.UnsupervisedLearning(E, data.dir = NULL)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{data.dir}{directory of SPRING files "edges.csv" and "categorical_coloring_data.json"}
}
\value{
Filtered markers where each marker must have at least ncells that express at least ncounts
}
\description{
Unsupervised main function
}
