% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.VinPlot}
\alias{CID.VinPlot}
\title{Makes probability distribution plots for genes across celltypes}
\usage{
CID.VinPlot(E, genes = c("MS4A1", "CD40", "NKG7"), lbls = NULL,
  do.print = FALSE, omit = NULL, filename = NULL)
}
\arguments{
\item{E}{expression matrix (genes by rows) where genes are rownames and columns are celltype labels (see ?CID.CellID)}

\item{genes}{a character vector of genes to plot.}
}
\value{
A probability distribution plot for the queried genes.
}
\description{
Makes probability distribution plots for genes across celltypes
}
