% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{geneconversion}
\alias{geneconversion}
\title{Convert gene ids}
\usage{
geneconversion(genestoconvert, from = "hugo")
}
\arguments{
\item{genestoconvert}{A character vector of N gene names (HUGO gene symbols)}

\item{from}{Default is hugo gene symbols. Anything besides "hugo" will assune ENTREZ gene symbols.}
}
\value{
A data frame containing the converted gene names
}
\description{
Convert gene ids
}
