% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{Generate_lbls}
\alias{Generate_lbls}
\title{Generate labels from classifications}
\usage{
Generate_lbls(cr, spring.dir = NULL, E = NULL, smooth = T)
}
\arguments{
\item{cr}{list returned by Signac function call.}

\item{spring.dir}{If using SPRING, directory to categorical_coloring_data.json. Default is NULL.}

\item{E}{a sparse gene (rows) by cell (column) matrix, or a Seurat object. Rows are HUGO symbols.}

\item{smooth}{if TRUE, smooths the cell type classifications. Default is TRUE.}
}
\value{
cell type labels (list) for each level of the hierarchy.
}
\description{
Generate labels from classifications
}
