% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{GenerateNetwork}
\alias{GenerateNetwork}
\title{Generate a network with Graphical LASSO}
\usage{
GenerateNetwork(E, gns_of_int, rho = c(0.3, 0.4, 0.5, 0.6), metadata)
}
\arguments{
\item{E}{Sparse expression matrix with rows genes and columns cells (see ?Signac)}

\item{gns_of_int}{<10,000 gene list of genes for network reconstruction}

\item{rho}{sparsity parameter for graphical LASSO}

\item{metadata}{data frame with metadata for each cell after signac classification}
}
\value{
network object
}
\description{
Generate a network with Graphical LASSO
}
