% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Genes_Of_Interest}
\alias{Genes_Of_Interest}
\title{Genes of interest for drug discovery / disease biology research}
\format{
A data frame with five columns
\describe{
  \item{Genes}{genes, identified by HUGO symbols}
  \item{CellPhoneDB}{0 if not in CellPhoneDB, 1 if gene is listed as a receptor in CellPhoneDB}
  \item{GWAS}{0 if not GWAS, 1 if GWAS in GWAS catalog}
  \item{PI_drugs}{0 if not in Priority Index paper, 1 if gene is listed as a drug target in priority index paper}
  \item{PI_GWAS}{0 if not in Priority Index GWAS list, 1 if gene is listed as GWAS in priority index paper}
  ...
}
}
\usage{
Genes_Of_Interest
}
\description{
3,304 genes curated from external sources
}
\keyword{datasets}
