% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{get_celltypes}
\alias{get_celltypes}
\title{Get main cell types from hierarchy}
\usage{
get_celltypes(R)
}
\arguments{
\item{R}{reference matrix}
}
\value{
a vector of cell type identities given the hierarchy
}
\description{
Get main cell types from hierarchy
}
