% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_colors}
\alias{get_colors}
\title{Get HEX colors}
\usage{
get_colors(P)
}
\arguments{
\item{P}{A character vector of cell type labels}
}
\value{
A vector of HEX colors for each category of cell type labels
}
\description{
Get HEX colors
}
\keyword{internal}
