% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{get_knn_graph2}
\alias{get_knn_graph2}
\title{Get KNN graph}
\usage{
get_knn_graph2(X, k = 4, np, genes_to_use)
}
\arguments{
\item{X}{Expression matrix with genes for rows, samples for columns, after V score filtering.}

\item{k}{KNN parameter. Default is 5.}

\item{np}{Number of PCs to build the KNN graph. Default is 50.}

\item{genes_to_use}{gene list for KNN graph building}
}
\description{
Get KNN graph
}
