% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{get_vscores_sparse}
\alias{get_vscores_sparse}
\title{Get V scores}
\usage{
get_vscores_sparse(
  E,
  min_mean = 0,
  nBins = 50,
  fit_percentile = 0.1,
  error_wt = 1
)
}
\arguments{
\item{E}{Expression matrix with genes for rows, samples for columns}

\item{min_mean}{Minimum mean gene expression. Default is zero.}

\item{nBins}{Number of bins for histogram binning. Default is 50.}

\item{fit_percentile}{Percentile for fitting. Default is 0.1.}

\item{error_wt}{Error for convergence of optimization. Default is 1.}
}
\value{
V scores for each gene in the expression matrix E.
}
\description{
Get V scores
}
