% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{GetAllMarkers}
\alias{GetAllMarkers}
\title{Differential Expression Analysis}
\usage{
GetAllMarkers(
  E,
  Samples = NULL,
  Disease = NULL,
  Identities,
  omit = c("none", "Empty", "Double200-0109&200-0611")
)
}
\arguments{
\item{E}{A gene by cell expression matrix}

\item{Samples}{A character vector of samples.}

\item{Disease}{A character vector of disease labels}

\item{Identities}{A character vector of cell type or cluster labels}

\item{omit}{Any samples to omit. Default is "none" and "Empty"}
}
\value{
a DEG table
}
\description{
Differential Expression Analysis
}
