% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{GetGeneFromNetwork}
\alias{GetGeneFromNetwork}
\title{Load gene from network h5 file}
\usage{
GetGeneFromNetwork(gene, filename = "network_sparse_genes.h5")
}
\arguments{
\item{gene}{A gene to query from the network}

\item{filename}{file name of network. Default is "network.hdf5"}
}
\value{
a saved network file
}
\description{
Load gene from network h5 file
}
