% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{GetMarkers}
\alias{GetMarkers}
\title{Differential Expression Analysis for reference dataset}
\usage{
GetMarkers(R)
}
\arguments{
\item{R}{A reference list as returned by data("Reference_sim")}
}
\value{
a list of DEG tables
}
\description{
Differential Expression Analysis for reference dataset
}
