% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{GetModels_HPCA}
\alias{GetModels_HPCA}
\title{Loads neural network models from GitHub}
\usage{
GetModels_HPCA()
}
\value{
list of 1,800 neural network models stacked to solve 18 classification problems.
}
\description{
\code{GetModels_HPCA} returns a list of neural network models that were trained with the HPCA training data.
The HPCA training data has 18 classification tasks with bootstrapped training data.
The training data are split into two distinct cellular phenotypes (i.e., immune and nonimmune).
\code{GetModels_HPCA} downloads pre-computed neural networks trained to classify cells for each task (n = 100 models trained for each tasks).
}
\examples{
\dontrun{
P = GetModels()
}
}
\seealso{
\code{\link{SignacFast}}, \code{\link{ModelGenerator}}
}
