% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{GetTrainingData_HPCA}
\alias{GetTrainingData_HPCA}
\title{Loads bootstrapped HPCA training data from GitHub}
\usage{
GetTrainingData_HPCA()
}
\value{
A list with two elements; 'Reference' are the training data, and 'genes' -- the union of all features present in the training data.
}
\description{
\code{GetTrainingData_HPCA} returns a list of bootstrapped HPCA training data.
The HPCA training data has 18 classification tasks, each split into two distinct cellular phenotypes (i.e., immune and nonimmune).
\code{GetTrainingData_HPCA} downloads bootstrapped training data to use for model-building.
}
\examples{
\dontrun{
P = GetTrainingData_HPCA()
}
}
