% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{getxCellGenes}
\alias{getxCellGenes}
\title{Get genes from xCell publication}
\usage{
getxCellGenes(dataset = NULL)
}
\arguments{
\item{dataset}{default is NULL}
}
\value{
a list of gene signatures
}
\description{
Get genes from xCell publication
}
