% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_markers}
\alias{immune_markers}
\title{The gene signatures used for immune cell type identification.}
\format{
A data.frame with 2058 rows and three columns.
\describe{
  \item{HUGO symbols}{genes, identified by HUGO symbols}
  \item{Cell population}{identified by common nomenclature}
  \item{Polarity}{+ for positive markers, - for negative markers}
  ...
}
}
\usage{
immune_markers
}
\description{
A group of 2058 markers used for immune cell type identification.
}
\keyword{datasets}
