% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{JackD}
\alias{JackD}
\title{Jackknife Differential Expression Analysis}
\usage{
JackD(E, Samples, Disease, Identities, do.par = F, num.cores = 1)
}
\arguments{
\item{E}{A gene by cell expression matrix}

\item{Samples}{A character vector of samples}

\item{Disease}{A character vector of disease labels}

\item{Identities}{A character vector of cell type or cluster labels}

\item{do.par}{if TRUE, code is run in parallel on all available cores minus one}

\item{num.cores}{optionally, user can hard set the number of cores to use}
}
\value{
a DEG table Jackknifed
}
\description{
Jackknife Differential Expression Analysis
}
