% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{KSoftImpute}
\alias{KSoftImpute}
\title{K soft imputation}
\usage{
KSoftImpute(E, dM = NULL, genes.to.use = NULL, do.save = F)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{data.dir}{directory of SPRING files "edges.csv" and "categorical_coloring_data.json"}

\item{genes}{a character vector of genes to impute}
}
\value{
Imputed values for every gene
}
\description{
K soft imputation
}
