% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{LoadNetwork}
\alias{LoadNetwork}
\title{Load gene from network h5 file}
\usage{
LoadNetwork(filename = "network_total.h5")
}
\arguments{
\item{filename}{file name of network. Default is "network.hdf5"}
}
\value{
a saved network file
}
\description{
Load gene from network h5 file
}
