% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{The gene signatures used for cell type identification.}
\format{
A data.frame with 2715 rows and three columns.
\describe{
  \item{HUGO symbols}{genes, identified by HUGO symbols}
  \item{Cell population}{identified by common nomenclature}
  \item{Polarity}{+ for positive markers, - for negative markers}
  ...
}
}
\usage{
markers
}
\description{
A group of 2715 markers used for cell type identification.
}
\keyword{datasets}
