% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{MASC}
\alias{MASC}
\title{Main function for mixed effect modeling}
\usage{
MASC(
  dataset,
  cluster,
  contrast,
  random_effects = NULL,
  fixed_effects = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset}{data frame of covariate, cell type, clustering or disease information}

\item{cluster}{celltypes returned by Signac or cluster identities}

\item{contrast}{Typically disease}

\item{random_effects}{User specified random effect variables in dataset}

\item{fixed_effects}{User specific fixed effects in dataset}

\item{verbose}{If TRUE, algorithm reports outputs}
}
\value{
mixed effect model results
}
\description{
Main function for mixed effect modeling
}
