% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{ModelGenerator}
\alias{ModelGenerator}
\title{Main function for generating an ensemble of machine learning models.}
\usage{
ModelGenerator(
  R,
  model.use = "nn",
  N = 1,
  num.cores = 1,
  verbose = T,
  hidden = 1,
  set.seed = T,
  seed = "42"
)
}
\arguments{
\item{R}{Reference dataset; user should do data("Reference_hpca") and then set R to Reference_hpca.}

\item{model.use}{Machine learning model to use. Default option is neural network. Can also be set to 'svm' or 'rf'.}

\item{N}{Number of machine learning models to train (for nn and svm). Default is 100.}

\item{num.cores}{Number of cores to use. Default is 1.}

\item{verbose}{if TRUE, code will report outputs. Default is TRUE.}

\item{hidden}{Number of hidden layers in the neural network. Default is 1.}

\item{set.seed}{If true, seed is set to ensure reproducibility of these results. Default is TRUE.}

\item{seed}{if set.seed is TRUE, seed is set to 42.}
}
\value{
models for predicting cell types based on reference dataset
}
\description{
Main function for generating an ensemble of machine learning models.
}
