% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Models_HPCA}
\alias{Models_HPCA}
\title{A list of machine learning models to use with SignacFast}
\format{
A list with 18 entries, each with 100 neural networks.
\describe{
  \item{List_element}{Each list element is an ensemble of neural networks}
  \item{genes}{genes used in the neural network model}
  ...
}
}
\usage{
Models_HPCA
}
\description{
List of neural network models (1,800) trained using the ModelGenerator function.
}
\keyword{datasets}
