% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{PathwayEnrichment}
\alias{PathwayEnrichment}
\title{Reactome Pathway Enrichment}
\usage{
PathwayEnrichment(genes, verbose = T)
}
\arguments{
\item{genes}{A character vector of gene symbols (ENTREZ gene symbols).}

\item{verbose}{Code speaks. Default is TRUE.}
}
\value{
Pathway enrichment results
}
\description{
Reactome Pathway Enrichment
}
