% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{PathwayEnrichmentPlot}
\alias{PathwayEnrichmentPlot}
\title{Reactome Pathway Enrichment plotting}
\usage{
PathwayEnrichmentPlot(
  X,
  N = 40,
  color = "red",
  qvalueCutoff = 0.05,
  adj.pvalCutoff = 0.05
)
}
\arguments{
\item{X}{results of ?PathwayEnrichment}

\item{N}{number of pathways to plot}

\item{color}{Color of dot plot. Default is red.}

\item{qvalueCutoff}{FDR cutoff. Default is 0.05.}

\item{adj.pvalCutoff}{Adjusted p-value cutoff. Default is 0.05.}
}
\value{
Pathway enrichment plot
}
\description{
Reactome Pathway Enrichment plotting
}
