% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{ProbabilityPlot}
\alias{ProbabilityPlot}
\title{Generate a probability plot}
\usage{
ProbabilityPlot(df)
}
\arguments{
\item{df}{one element of the list returned by Signac function call (see ?Signac)}
}
\value{
ggplot object
}
\description{
Generate a probability plot
}
