% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_learned}
\alias{R_learned}
\title{A single reference data set used for identifying CD56+ NK cells}
\format{
A data frame with 32 columns (features + cell type labels) and 2,000 bootstrapped cells
\describe{
  \item{celltypes}{Labels for classification (celltypes)}
  ...
}
}
\usage{
R_learned
}
\description{
For use with Signac_Solo
}
\keyword{datasets}
