% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reference_sim}
\alias{Reference_sim}
\title{Reference expression matrix bootstrapped for cell type / cell state signatures.}
\format{
A list with two entries
\describe{
  \item{genes}{genes, identified by HUGO symbols}
  \item{Reference}{different training datasets for binary cell type classification}
  ...
}
}
\usage{
Reference_sim
}
\description{
A reference dataset with 1,000 pure cell types
}
\keyword{datasets}
