% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{runningquantile}
\alias{runningquantile}
\title{Compute running quantile}
\usage{
runningquantile(x, y, p, nBins)
}
\arguments{
\item{x}{log of mean gene expression.}

\item{y}{log of fano factor divided by mean gene expression.}

\item{p}{fit_percentile as defined in get_vscores_sparse}

\item{nBins}{number of bins for histogram binning}
}
\description{
Compute running quantile
}
