% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S}
\alias{S}
\title{Reference expression matrix for cell type / cell state signatures.}
\format{A matrix with 9555 rows and 224 columns
\describe{
  \item{genes}{genes, identified by HUGO symbols}
  \item{sorted cells}{identified by common nomenclature}
  ...
}}
\usage{
S
}
\description{
A reference dataset with 9555 genes and 224 experiments (mostly sorted bulk RNA seq)
}
\keyword{datasets}
