% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{SaveCountsToH5}
\alias{SaveCountsToH5}
\title{Write results to a folder}
\usage{
SaveCountsToH5(D, data.dir, genome = "GRCh38")
}
\arguments{
\item{D}{A list of count matrices}

\item{data.dir}{directory (will be created if it does not exist) where results are saved}

\item{genome}{default is GRCh38.}
}
\value{
matrix.h5 file, where each is background corrected
}
\description{
Write results to a folder
}
