% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{SaveNetworksH5}
\alias{SaveNetworksH5}
\title{Save network h5 files}
\usage{
SaveNetworksH5(A, data.dir)
}
\arguments{
\item{A}{an adjacency matrix with row names and columns names genes (gene by gene)}

\item{data.dir}{directory where the networks are saved}
}
\value{
a saved network file
}
\description{
Save network h5 files
}
