% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{ShowHierarchy}
\alias{ShowHierarchy}
\title{Show hierarchical structure}
\usage{
ShowHierarchy(R)
}
\arguments{
\item{R}{Reference dataset}
}
\value{
annotations by hierarchy
}
\description{
Show hierarchical structure
}
