% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{SignacLearn}
\alias{SignacLearn}
\title{Main function for learning}
\usage{
SignacLearn(
  E,
  learned_types,
  labels,
  size = 1000,
  impute = T,
  spring.dir = NULL,
  logfc.threshold = 0.25,
  p.val.adj = 0.05,
  verbose = T
)
}
\arguments{
\item{E}{a sparse gene (rows) by cell (column) matrix. Rows are HUGO symbols.}

\item{learned_types}{Types for learning}

\item{labels}{cell type labels for the columns of E.}

\item{size}{Number of bootstrapped samples for machine learning. Default is 1,000.}

\item{impute}{if TRUE, performs imputation prior to bootstrapping. Default is TRUE.}

\item{spring.dir}{if using SPRING, directory to categorical_coloring_data.json. Default is NULL.}

\item{logfc.threshold}{Cutoff for feature selection. Default is 0.25.}

\item{p.val.adj}{Cutoff for feature selection. Default is 0.05.}

\item{verbose}{if TRUE, code speaks. Default is TRUE.}
}
\value{
training data frame to be added to a new reference training set
}
\description{
Main function for learning
}
