% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{sparse_var}
\alias{sparse_var}
\title{Compute variance for sparse matrices.}
\usage{
sparse_var(E)
}
\arguments{
\item{E}{Expression matrix with genes for rows, samples for columns}
}
\description{
Compute variance for sparse matrices.
}
