% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.append}
\alias{CID.append}
\title{Calculates Signac scores}
\usage{
CID.append(expression, featuresG, sorted)
}
\arguments{
\item{expression}{An expression matrix with features (genes) in rows and samples (cells) in columns.}

\item{featuresG}{A list of features, where each list entry is a character vector of markers.}

\item{sorted}{see ?CID.CellID}
}
\value{
A matrix of CellID scores; features by cells.
}
\description{
Calculates Signac scores
}
