% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.CellID}
\alias{CID.CellID}
\title{Main function}
\usage{
CID.CellID(E, normalize = F, pval = 0.05, data.dir = NULL,
  entropy = T, louvain = T, omit = NULL, sorted = F)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{normalize}{Normalizes count matrix to the mean library size. Default is TRUE.}

\item{pval}{p-value cutoff for feature selection, as described in the manuscript + markdown file. Default is pval = 0.05.}

\item{data.dir}{directory of SPRING files "edges.csv" and "categorical_coloring_data.json"}

\item{entropy}{cells amended to high entropy labels with respect to their neighbors in the KNN graph are appended "Other" if entropy = TRUE. Default is TRUE.}

\item{louvain}{Louvain community detection is performed, and then used together with Shannon entropy to detect potential novel cell types / states. Default is TRUE.}

\item{omit}{Force remove specific cell types / states with omit. Default is NULL.}

\item{sorted}{If cells are expected to be pure or mostly homogeneous (e.g., by FACs sorting), set sorted = TRUE. Default is FALSE.}
}
\value{
Filtered markers where each marker must have at least ncells that express at least ncounts
}
\description{
Main function
}
