% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.CellID}
\alias{CID.CellID}
\title{Main function}
\usage{
CID.CellID(E, f = NULL, pval = 0.1, deep_dive = TRUE, edges = NULL,
  entropy = FALSE)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols (see ?CID.geneconversion), or a list of such matrices, see ?CID.BatchMode.}

\item{f}{Parameter for classification smoothing. Default is f = NULL if edges = NULL, else default is f = 0.5 automatically.}

\item{pval}{p-value cutoff for feature selection, as described in the manuscript + markdown file. Default is pval = 0.1.}

\item{deep_dive}{Boolean, T will assign cell types and then cell states, F will only assign cell types. Default is deep_dive = T.}

\item{edges}{Data-frame containing the edgelist for graph edges for smoothing, OR a directory where "edges.csv" is located, OR a character vector / list for batch mode. Default is edges = NULL.}
}
\value{
Filtered markers where each marker must have at least ncells that express at least ncounts
}
\description{
Main function
}
