% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.entropy}
\alias{CID.entropy}
\title{Entropy}
\usage{
CID.entropy(ac, edges)
}
\arguments{
\item{edges}{A data frame or matrix with edges}
}
\value{
The entropy for each node in the network
}
\description{
Entropy
}
