% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.entropy}
\alias{CID.entropy}
\title{Entropy}
\usage{
CID.entropy(ac, dM)
}
\arguments{
\item{ac}{A character vector of cell type labels}

\item{dM}{The distance matrix, see ?CID.GetDistMat}
}
\description{
Entropy
}
