% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.GetDistMat}
\alias{CID.GetDistMat}
\title{Get distance matrix}
\usage{
CID.GetDistMat(data.dir, n = 4)
}
\arguments{
\item{data.dir}{directory where edges.csv is located}

\item{n}{maximum network distance to subtend (n neighbors)}
}
\value{
adjacency matrices for distances < n
}
\description{
Get distance matrix
}
