% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.GetEdges}
\alias{CID.GetEdges}
\title{Get edges that are either pre-computed, or generate new edges}
\usage{
CID.GetEdges(...)
}
\value{
edges for cell-cell similarity network
}
\description{
Get edges that are either pre-computed, or generate new edges
}
