% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.Impute}
\alias{CID.Impute}
\title{Imputation wrapper}
\usage{
CID.Impute(E, data.dir = NULL, do.par = TRUE)
}
\arguments{
\item{E}{A gene-by-sample count matrix (sparse matrix, matrix, or data.frame) with genes identified by their HUGO symbols.}

\item{data.dir}{directory for saving "matrix_saver_imputed.mtx" for future loading.}

\item{do.par}{Boolean. If true, imputation is performed in parallel on half of the machines available cores. Default = FALSE.}
}
\value{
imputed expression matrix with only marker genes in rows.
}
\description{
Imputation wrapper
}
