% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CID.IsUnique}
\alias{CID.IsUnique}
\title{Extracts unique elements}
\usage{
CID.IsUnique(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
boolean, unique elements are TRUE
}
\description{
\code{CID.IsUnique} returns a Boolean for the unique elements of a vector.
}
\examples{
# generate a dummy variable
dummy = c("A", "A", "B", "C")

# get only unique elements
logik = CID.IsUnique(dummy) 
dummy[logik]
}
