% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CID.LoadData}
\alias{CID.LoadData}
\title{Load data file from directory}
\usage{
CID.LoadData(data.dir, mfn = "matrix.mtx")
}
\arguments{
\item{data.dir}{Directory containing matrix.mtx and genes.txt.}

\item{mfn}{file name; default is 'matrix.mtx'}
}
\value{
A sparse matrix with rownames equivalent to the names in genes.txt
}
\description{
Loads 'matrix.mtx' and 'genes.txt' files from a directory.
}
\examples{
\dontrun{
# Loads data from SPRING

# dir is your path to the "categorical_coloring_data.json" file
dir = "./FullDataset_v1"

# load expression data
E = CID.LoadData(data.dir = dir)
}
}
