% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{CID.smooth}
\alias{CID.smooth}
\title{Smoothing function}
\usage{
CID.smooth(Q, f, edges)
}
\arguments{
\item{Q}{List containing a character vector where each element is a cell type or cell state assignment}

\item{f}{Smoothing parameter}

\item{edges}{Location where knn edges are stored}
}
\value{
Smoothed cell type or cell state assignments
}
\description{
Smoothing function
}
