% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{Generate_lbls}
\alias{Generate_lbls}
\title{Generate labels from classifications}
\usage{
Generate_lbls(
  cr,
  spring.dir = NULL,
  E = NULL,
  smooth = T,
  new_populations = NULL,
  new_categories = NULL,
  min.cells = 10
)
}
\arguments{
\item{cr}{list returned by Signac function call.}

\item{spring.dir}{If using SPRING, directory to categorical_coloring_data.json. Default is NULL.}

\item{E}{a sparse gene (rows) by cell (column) matrix, or a Seurat object. Rows are HUGO symbols.}

\item{smooth}{if TRUE, smooths the cell type classifications. Default is TRUE.}

\item{new_populations}{Character vector specifying any new cell types that were learned by Signac. Default is NULL.}

\item{new_categories}{If new_populations are set to a cell type, new_category is a corresponding character vector indicating the population that the new population belongs to. Default is NULL.}

\item{min.cells}{If desired, any cell population with equal to or less than N cells is set to "Unclassified." Default is 10 cells.}
}
\value{
cell type labels (list) for each level of the hierarchy.
}
\description{
Generate labels from classifications
}
