% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CID.R
\name{Signac_Solo}
\alias{Signac_Solo}
\title{Main function for solo-decision classification}
\usage{
Signac_Solo(
  E,
  R,
  spring.dir = NULL,
  model.use = "nn",
  N = 25,
  num.cores = 1,
  threshold = 0.5,
  smooth = T,
  impute = T,
  verbose = T,
  do.normalize = T,
  probability = F,
  hidden = 1
)
}
\arguments{
\item{E}{a sparse gene (rows) by cell (column) matrix, or a Seurat object. Rows are HUGO symbols.}

\item{R}{Reference dataset; user should do data("Reference_sim") and then set R to Refernence_sim.}

\item{spring.dir}{If using SPRING, directory to categorical_coloring_data.json. Default is NULL.}

\item{model.use}{Machine learning model to use. Default option is neural network. Can also be set to 'svm' or 'rf'.}

\item{N}{Number of machine learning models to train (for nn and svm). Default is 25.}

\item{num.cores}{Number of cores to use. Default is 1.}

\item{threshold}{Probability threshold for assigning cells to "Unclassified." Default is 0.5.}

\item{smooth}{if TRUE, smooths the cell type classifications. Default is TRUE.}

\item{impute}{if TRUE, gene expression values are imputed prior to cell type classification. Default is TRUE.}

\item{verbose}{if TRUE, code will report outputs. Default is TRUE.}

\item{do.normalize}{if TRUE, cells are normalized to the mean library size. Default is TRUE.}

\item{probability}{if TRUE, returns the probability associated with each cell type label. Default is TRUE.}

\item{hidden}{Number of hidden layers in the neural network. Default is 1.}
}
\value{
annotations
}
\description{
Main function for solo-decision classification
}
